% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections_get.R
\name{collections_get}
\alias{collections_get}
\title{Retrieve data about a specific collection.}
\usage{
collections_get(page_id)
}
\arguments{
\item{page_id}{Page id of active page in Domo.}
}
\value{
List of raw output from Domo API
}
\description{
Retrieve data about a specific collection.
}
\examples{
collections_get()
}
