% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections_update.R
\name{collections_update}
\alias{collections_update}
\title{Update a collection spec}
\usage{
collections_update(page_id, collection_id, collection_def)
}
\arguments{
\item{page_id}{Page id from Domo}

\item{collection_id}{Collection id from Domo}

\item{collection_def}{Updated page definition}
}
\value{
Some output from Domo
}
\description{
Update a collection spec
}
\examples{
page_update()
}
