% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdomo.R
\docType{class}
\name{Domo-class}
\alias{Domo-class}
\alias{Domo}
\title{Reference class containing functionality to interact with Domo.}
\description{
Inherits fields from DomoUtilities
}
\details{
All methods documented separately via standard documentation methods.
}
\section{Methods}{

\describe{
\item{\code{ds_create(df_up, name, description = "", update_method = "REPLACE")}}{Create a new data set.}

\item{\code{ds_delete(ds, prompt_before_delete = TRUE)}}{Delete a data set.}

\item{\code{ds_get(ds, r_friendly_names = TRUE, ...)}}{Download data from Domo into a dataframe (tibble).}

\item{\code{ds_list(limit = 0, offset = 0, df_output = TRUE)}}{List all data sets}

\item{\code{ds_meta(...)}}{Get all meta data related to a data set.}

\item{\code{ds_query(ds, query, return_data = TRUE)}}{Evaluate a query against a data set.}

\item{\code{ds_update(ds_id, df_up)}}{Update an existing data set.}
}}

