% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_delete.R
\name{ds_delete}
\alias{ds_delete}
\title{Delete a given data set.}
\usage{
ds_delete(ds, prompt_before_delete = TRUE)
}
\arguments{
\item{ds}{Data set id from Domo.}
}
\value{
NULL value
}
\description{
Delete a given data set.
}
\examples{
ds_delete('32e44f16-c76b-4de3-ac0e-1b7d2ad2f15d')
}
