% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_get.R
\name{ds_get}
\alias{ds_get}
\title{Download a given data set.}
\usage{
ds_get(ds, r_friendly_names = TRUE)
}
\arguments{
\item{ds}{Data set id from Domo.}
}
\value{
Dataframe of data set from Domo.
}
\description{
Download a given data set.
}
\examples{
ds_get('32e44f16-c76b-4de3-ac0e-1b7d2ad2f15d')
}
