% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_import.R
\name{ds_import}
\alias{ds_import}
\title{Upload a given data set.}
\usage{
ds_import(ds_id, df_up, update_method = "REPLACE")
}
\arguments{
\item{ds_id}{Data set id (from url in Domo).}

\item{df_up}{Dataframe to upload.}

\item{update_method}{Optional choose the appropriate updateMethod when importing data via the API. Available choices are either to APPEND or REPLACE.}
}
\value{
I'm sure something comes back...
}
\description{
Upload a given data set.
}
\examples{
ds_import('some_id',my_new_df)
}
