% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_list.R
\name{ds_list}
\alias{ds_list}
\title{List Data Sets}
\usage{
ds_list(limit = 0, offset = 0, df_output = TRUE)
}
\arguments{
\item{limit}{Number of results to return. Values < 1 will return all data sets.}

\item{offset}{Start location of search.}

\item{df_output}{TRUE/FALSE To return a data frame or a list of lists.}
}
\value{
List of data sets in either data frame or list of lists
}
\description{
List Data Sets
}
\examples{
ds_get()
}
