% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_meta.R
\name{ds_meta}
\alias{ds_meta}
\title{Retrieve meta data for a given data set.}
\usage{
ds_meta(ds)
}
\arguments{
\item{ds}{Data set id from Domo.}
}
\value{
List object containing data set meta data.
}
\description{
Retrieve meta data for a given data set.
}
\examples{
ds_meta('32e44f16-c76b-4de3-ac0e-1b7d2ad2f15d')
}
