% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_rename.R
\name{ds_rename}
\alias{ds_rename}
\title{Rename a given data set}
\usage{
ds_rename(ds, new_name, new_description = "")
}
\arguments{
\item{ds}{Data set id from Domo.}

\item{new_name}{The new name of the data set}

\item{new_description}{The new description of the data set (optional)}
}
\value{
NULL value
}
\description{
Rename a given data set
}
\examples{
ds_rename('32e44f16-c76b-4de3-ac0e-1b7d2ad2f15d','NEW NAME',description='THIS IS A DESCRIPTION')
}
