% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_update.R
\name{ds_update}
\alias{ds_update}
\title{Upload a given data set.}
\usage{
ds_update(ds_id, df_up)
}
\arguments{
\item{ds_id}{Data set id (from url in Domo).}

\item{df_up}{Dataframe to upload.}
}
\value{
I'm sure something comes back...
}
\description{
Upload a given data set.
}
\examples{
ds_import('some_id',my_new_df)
}
