% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_access.R
\name{get_access}
\alias{get_access}
\title{Get access token environment variable for data access.}
\usage{
get_access(type = "data")
}
\value{
Sucess or failure (1/0). An access token for data acess and one for user access will be stored in the package environment.
}
\description{
Based on given client id and secret from developer.domo.com, get access token for data acess and store in environment.
We assume the given OAuth information has at least data access.
}
\examples{
get_access()
}
