% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_domain.R
\name{get_domain}
\alias{get_domain}
\title{Get domain, some instances are on beta}
\usage{
get_domain(type = "data")
}
\value{
Returns the correct domain based on client id and secret
}
\description{
Need to get the right one, right?
}
\examples{
get_domain()
}
