% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_add_users.R
\name{groups_add_users}
\alias{groups_add_users}
\title{Return group definition.}
\usage{
groups_add_users(group_id, users)
}
\arguments{
\item{group_id}{Group id.}

\item{users}{List or array of users to add.}
}
\value{
Group properties
}
\description{
Return group definition.
}
\examples{
groups_get()
}
