% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_create.R
\name{groups_create}
\alias{groups_create}
\title{Create a new group. Add users if given.}
\usage{
groups_create(group_name, users = -1, active = "true")
}
\arguments{
\item{group_name}{Group id.}

\item{users}{List or array of users to remove.}
}
\value{
Group properties
}
\description{
Create a new group. Add users if given.
}
\examples{
groups_create()
}
