% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_delete.R
\name{groups_delete}
\alias{groups_delete}
\title{Delete group.}
\usage{
groups_delete(group_id)
}
\arguments{
\item{group_id}{Group id.}
}
\value{
Group properties
}
\description{
Delete group.
}
\examples{
groups_delete()
}
