% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_get.R
\name{groups_get}
\alias{groups_get}
\title{Return group definition.}
\usage{
groups_get(group_id)
}
\arguments{
\item{group_id}{Group id.}
}
\value{
Group properties
}
\description{
Return group definition.
}
\examples{
groups_get()
}
