% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_list.R
\name{groups_list}
\alias{groups_list}
\title{List all groups in Domo instance.}
\usage{
groups_list(df_output = TRUE)
}
\value{
List of all groups in Domo.
}
\description{
List all groups in Domo instance.
}
\examples{
groups_list()
}
