% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_list_users.R
\name{groups_list_users}
\alias{groups_list_users}
\title{Return group definition.}
\usage{
groups_list_users(group_id)
}
\arguments{
\item{group_id}{Group id.}
}
\value{
Group properties
}
\description{
Return group definition.
}
\examples{
groups_list_users()
}
