% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups_remove_users.R
\name{groups_remove_users}
\alias{groups_remove_users}
\title{Remove users from group.}
\usage{
groups_remove_users(group_id, users)
}
\arguments{
\item{group_id}{Group id.}

\item{users}{List or array of users to remove.}
}
\value{
Group properties
}
\description{
Remove users from group.
}
\examples{
groups_remove_users()
}
