% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_create.R
\name{page_create}
\alias{page_create}
\title{Create a page}
\usage{
page_create(page_def)
}
\arguments{
\item{page_def}{List object with page definition}
}
\value{
List of pages
}
\description{
Download list of pages
}
\examples{
page_create()
}
