% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_get.R
\name{page_get}
\alias{page_get}
\title{Download information regarding a specific Domo page.}
\usage{
page_get(page_id)
}
\arguments{
\item{page_id}{Page id of active page in Domo.}
}
\value{
List of raw output from Domo API
}
\description{
Download information regarding a specific Domo page.
}
\references{
API Documentation: \url{https://developer.domo.com/docs/page-api-reference/page#Retrieve\%20a\%20page}
}