% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_update.R
\name{page_update}
\alias{page_update}
\title{Update a page spec}
\usage{
page_update(page_id, page_def)
}
\arguments{
\item{page_id}{Page id from Domo}

\item{page_def}{Updated page definition}
}
\value{
Some output from Domo
}
\description{
Update a page spec
}
\references{
API Documentation: \url{https://developer.domo.com/docs/page-api-reference/page#Update\%20a\%20page}
}