% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdp_create.R
\name{pdp_create}
\alias{pdp_create}
\title{List all PDP policies defined on a given dataset.}
\usage{
pdp_create(ds, policy_def)
}
\arguments{
\item{ds}{Data set id from Domo.}

\item{policy_def}{List object with parameters as given on Git.}
}
\value{
List with all policy properties plus the id assigned by Domo.
}
\description{
List all PDP policies on a given dataset.
}
\examples{
pdp_create('32e44f16-c76b-4de3-ac0e-1b7d2ad2f15d',my_new_policy)
}
