% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdp_delete.R
\name{pdp_delete}
\alias{pdp_delete}
\title{Delete a specific policy from the given dataset.}
\usage{
pdp_delete(ds, policy)
}
\arguments{
\item{ds}{Data set id from Domo.}

\item{policy}{Id of existing policy.}
}
\value{
Success message (maybe??)
}
\description{
Delete a specific policy from the given dataset.
}
\examples{
pdp_delete('32e44f16-c76b-4de3-ac0e-1b7d2ad2f15d',7)
}
