% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdp_enable.R
\name{pdp_enable}
\alias{pdp_enable}
\title{Enable PDP on a data set}
\usage{
pdp_enable(ds, new_state)
}
\arguments{
\item{ds}{Data set id from Domo.}

\item{new_state}{Either TRUE or FALSE depending on what you want done.}
}
\value{
List with all data set properties (including whether or not PDP is enabled).
}
\description{
This is a special case of the ds_update function just to toggle PDP on and off. The ds_update function could also be used for this purpose.
}
\references{
API Documentation: \url{https://developer.domo.com/docs/dataset-api-reference/dataset#Update\%20a\%20DataSet}
}