% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdp_update.R
\name{pdp_update}
\alias{pdp_update}
\title{Update a single PDP policy}
\usage{
pdp_update(ds, policy, policy_def)
}
\arguments{
\item{ds}{Data set id from Domo.}

\item{policy}{Id of existing policy.}

\item{policy_def}{List defining updates to the existing policy.}
}
\value{
List of new policy definition.
}
\description{
To update an individual PDP policy, you'll need to specify all required fields for a PDP policy. Do this via a list object, the function will convert to JSON automatically.
}
\references{
API Documentation: \url{https://developer.domo.com/docs/dataset-api-reference/dataset#Update\%20a\%20Personalized\%20Data\%20Permission\%20(PDP)\%20policy}
}