% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_access.R
\name{set_access}
\alias{set_access}
\title{Set access token environment variable for data access.}
\usage{
set_access(my_cid, my_secret, type = "data")
}
\value{
Sucess or failure (1/0). An access token for data acess and one for user access will be stored in the package environment.
}
\description{
Based on given client id and secret from developer.domo.com, get access token for data acess and store in environment.
We assume the given OAuth information has at least data access.
}
\examples{
set_access()
}
