% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_search.R
\name{stream_search}
\alias{stream_search}
\title{Search for a stream based on a data set id}
\usage{
stream_search(ds)
}
\arguments{
\item{ds}{Data set id from Domo.}
}
\value{
List object containing search results
}
\description{
search for a stream based on a data set id
}
\examples{
stream_search('32e44f16-c76b-4de3-ac0e-1b7d2ad2f15d')
}
