% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_update.R
\name{stream_update}
\alias{stream_update}
\title{Update a stream based on a stream id obtained from stream search}
\usage{
stream_update(stream_id, updateMethod)
}
\arguments{
\item{stream_id}{stream_id from stream search.}

\item{updateMethod}{The data import behavior REPLACE/APPEND}
}
\value{
List object containing update results
}
\description{
Update a stream based on a stream id obtained from stream search
}
\examples{
stream_update('42')
}
