% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_add.R
\name{users_add}
\alias{users_add}
\title{Add a user or users}
\usage{
users_add(x_name, x_email, x_role, x_sendInvite = "false")
}
\arguments{
\item{x_name}{User's name}

\item{x_email}{Users's email address}

\item{x_role}{Users's Domo role (one of Admin, Priveleged, Participant)}

\item{x_sendInvite}{'true' if you want an email sent or 'false' if not. Defaults to 'false'}
}
\value{
Some output from Domo
}
\description{
Add a user or users
}
\examples{
users_add()
}
