\name{users_get}
\alias{users_get}
\title{Download information regarding a specific Domo user.}
\usage{
users_get(user_id)
}
\arguments{
\item{user_id}{User id for specific user account.}
}
\value{
List of raw output from Domo API.
}
\description{
Full user object from Domo's API.
}
\references{
API Documentation: \url{https://developer.domo.com/docs/users-api-reference/users-2#Retrieve\%20a\%20user}
}