% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_update.R
\name{users_update}
\alias{users_update}
\title{Update a user}
\usage{
users_update(user_id, user_def)
}
\arguments{
\item{user_id}{User id from Domo}

\item{user_def}{Updated user definition}
}
\value{
Updated user object from Domo.
}
\description{
Update a user object. Tip: get the existing user object first, then modify elements of that object for the update.
}
\references{
API Documentation: \url{https://developer.domo.com/docs/users-api-reference/users-2#List\%20users}
}