% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_create.R
\name{ds_create}
\alias{ds_create}
\title{Download a given data set.}
\usage{
ds_create(df_up, name, description = "", update_method = "REPLACE",
  other_props = -1)
}
\arguments{
\item{df_up}{Dataframe to upload.}

\item{name}{Data set name}

\item{description}{Optional data set description.}

\item{update_method}{Optional choose the appropriate updateMethod when importing data via the API. Available choices are either to APPEND or REPLACE.}

\item{other_props}{List of addtional properties to send. Currently non-functional}
}
\value{
I'm sure something comes back...
}
\description{
Download a given data set.
}
\examples{
ds_create(my_new_df,'Some name')
}
