% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_create.R
\name{page_create}
\alias{page_create}
\title{Create a page}
\usage{
page_create(page_def)
}
\arguments{
\item{page_def}{List object with page definition. Link in references below for list specification.}
}
\value{
List of pages
}
\description{
This function will create a page in Domo. See references for more information.
}
\references{
API Documentation: \url{https://developer.domo.com/docs/page-api-reference/page#Create\%20a\%20page}
}