% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_list.R
\name{page_list}
\alias{page_list}
\title{Download list of pages}
\usage{
page_list(limit = 500, offset = 0)
}
\arguments{
\item{limit}{Limit number of results (default = max = 500)}

\item{offset}{Offset when paging through results}
}
\value{
List of pages
}
\description{
Download list of pages
}
\examples{
page_list()
}
