% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdp_delete.R
\name{pdp_delete}
\alias{pdp_delete}
\title{Delete a specific policy from the given dataset.}
\usage{
pdp_delete(ds, policy)
}
\arguments{
\item{ds}{Data set id from Domo.}

\item{policy}{Id of existing policy.}
}
\value{
Success message (maybe??)
}
\description{
Delete a specific policy from the given dataset.

NOTE: This is destructive. Once the PDP policy is deleted, it can't be restored w/o creating a new policy from scratch. It is recommended to backup policies before deleting.
}
\references{
API Documentation: \url{https://developer.domo.com/docs/dataset-api-reference/dataset#Delete\%20a\%20Personalized\%20Data\%20Permission\%20(PDP)\%20policy}
}