% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdp_enable.R
\name{pdp_enable}
\alias{pdp_enable}
\title{List all PDP policies defined on a given dataset.}
\usage{
pdp_enable(ds, new_state)
}
\arguments{
\item{ds}{Data set id from Domo.}

\item{new_state}{Either TRUE or FALSE depending on what you want done.}
}
\value{
List with all data set properties (I think).
}
\description{
List all PDP policies on a given dataset.
}
\examples{
pdp_enable('32e44f16-c76b-4de3-ac0e-1b7d2ad2f15d','TRUE')
}
