% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdp_list.R
\name{pdp_list}
\alias{pdp_list}
\title{List all PDP policies defined on a given dataset.}
\usage{
pdp_list(ds)
}
\arguments{
\item{ds}{Data set id from Domo.}
}
\value{
List with all policies defined on a data set.
}
\description{
List all PDP policies on a given dataset.
}
\examples{
pdp_list('32e44f16-c76b-4de3-ac0e-1b7d2ad2f15d')
}
