% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdp_update.R
\name{pdp_update}
\alias{pdp_update}
\title{List all PDP policies defined on a given dataset.}
\usage{
pdp_update(ds, policy, policy_def)
}
\arguments{
\item{ds}{Data set id from Domo.}

\item{policy}{Id of existing policy.}

\item{policy_def}{List defining updates to the existing policy.}
}
\value{
List of new policy definition.
}
\description{
List all PDP policies on a given dataset.
}
\examples{
pdp_update('32e44f16-c76b-4de3-ac0e-1b7d2ad2f15d',some,stuff)
}
