% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_get.R
\name{users_get}
\alias{users_get}
\title{Download information regarding a specific Domo user.}
\usage{
users_get(user_id)
}
\arguments{
\item{user_id}{Page id of active page in Domo.}
}
\value{
List of raw output from Domo API
}
\description{
Download information regarding a specific Domo user.
}
\examples{
users_get()
}
