% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_list.R
\name{users_list}
\alias{users_list}
\title{List all users in Domo instance.}
\usage{
users_list(df_output = TRUE, batch_size = 500)
}
\value{
List of all users in Domo.
}
\description{
List all users in Domo instance.
}
\examples{
users_list()
}
