% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_update.R
\name{users_update}
\alias{users_update}
\title{Update a user}
\usage{
users_update(user_id, user_def)
}
\arguments{
\item{user_id}{Page id from Domo}

\item{user_def}{Updated page definition}
}
\value{
Some output from Domo
}
\description{
Update a user
}
\examples{
user_update()
}
